# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Workforce Analytics & IT Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List enrollment tokens
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnrollmentTokenItem"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Mint a new enrollment token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentTokenItem"
  /tokens/{id}/revoke:
    post:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke an enrollment token
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentTokenItem"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    EnrollmentTokenItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        label:
          type: [string, "null"]
        createdById:
          type: [string, "null"]
          format: uuid
        assignedUserId:
          type: [string, "null"]
          format: uuid
        maxUses:
          type: integer
        useCount:
          type: integer
        expiresAt:
          type: [string, "null"]
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        enrolledDevices:
          type: array
          items:
            $ref: "#/components/schemas/EnrolledDeviceRef"
      required: [id, token, maxUses, useCount, createdAt, enrolledDevices]
    EnrolledDeviceRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        systemName:
          type: string
      required: [id, systemName]
    CreateTokenRequest:
      type: object
      properties:
        label:
          type: string
        maxUses:
          type: integer
        expiresDays:
          type: integer
...